+++ commercetools-sdk-java-api-customtypes-generator/src/main/kotlin/de/akii/commercetools/api/customtypes/generator/model/product/variantAttributes.kt
@@ -89,7 +89,7 @@ private fun typeNameForAttributeType(attributeType: AttributeType, isRequired: B
         is AttributeLocalizableTextType -> LocalizedString::class.asTypeName()
         is AttributeEnumType -> AttributePlainEnumValue::class.asTypeName()
         is AttributeLocalizedEnumValue -> AttributeLocalizedEnumValue::class.asTypeName()
-        is AttributeNumberType -> Int::class.asTypeName()
+        is AttributeNumberType -> Double::class.asTypeName()
         is AttributeMoneyType -> TypedMoney::class.asTypeName()
         is AttributeDateType -> LocalDate::class.asTypeName()
         is AttributeTimeType -> LocalTime::class.asTypeName()
